New Rule: InsufficientStringbufferDeclaration
See: http://sourceforge.net/forum/forum.php?thread_id=1438119&forum_id=188194

Rule finds StringBuffers which may be incorrectly pre-sized. It takes a worst case scenario from any if/switch blocks, and adds the lengths of all string/primative literals that are appended into the StringBuffer. It compares this against the length the constructor was given, if any. Default for a StringBuffer is 16 characters, and a buffer allocated with a String gets that String's length+16 characters.

If the constructor is created in a way that the length can not be calculated, the buffer is skipped, assuming the developer knew what they were doing.

The savings here is needless re-sizing.  If the string buffer is improperly sized, when your appends pass the current length, if forces new memory to be allocated in order to resize the buffer.


git-svn-id: https://pmd.svn.sourceforge.net/svnroot/pmd/trunk@4265 51baf565-9d33-0410-a72c-fc3788e3496d