Fixed bug 1031966
Re-Implemented CloneMethodMustImplementCloneable as a typeresolution rule. This rule can now detect super classes/interfaces which are cloneable.

The JUnit test was a bit tricky - I've added the interface as an actual concrete interface inside the xml package, and reference it from the xml test definitions. The JUnit works, this sort of design may be needed for other type resolution rules


git-svn-id: https://pmd.svn.sourceforge.net/svnroot/pmd/trunk@4875 51baf565-9d33-0410-a72c-fc3788e3496d