Applied patch from : 1751053 - New XSL a la Checkstyle/Maven.
I removed a couple of System.out that i had forgotten about. 
I also tried to integrate this new Render into the PMDAntTask.

AntTask will now recognize this formatter:

<formatter      type="betterhtml" .../>

However, it failed very weirdly :

build.xml:81: Can't file XSLT sheet :etc/pmd-nicerhtml.xsl

Which is kind of weird as i use getClass().getResourceAsStream() to find the xsl, which is in the jar:

$ unzip -t pmd-4.0.jar | grep -e 'etc/pmd-n'
    testing: etc/pmd-nicerhtml.xsl    OK

I believe it is a classloarder problem, but i maybe wrong. I'll try to deceipher this later...


git-svn-id: https://pmd.svn.sourceforge.net/svnroot/pmd/trunk@5493 51baf565-9d33-0410-a72c-fc3788e3496d