Add support for adding Rules to a RuleSet by reference.  Useful for IDE plugins, which intend to allow defining custom RuleSets using existing Rules without the current copy semantics.  Essentially, this lets us more easily do programmatically what we can already do in a hand-written XML.
For this to work, a Rule/RuleSet being added by Reference must have a file name associated with it. This information will be associated with the RuleSet when created by the RuleSetFactory using the APIs which provide file names/resources.  If a raw InputStream form is used, then the file name will need to be manually set.

git-svn-id: https://pmd.svn.sourceforge.net/svnroot/pmd/trunk@5777 51baf565-9d33-0410-a72c-fc3788e3496d