Vast expansion of Type Resolution scope on the AST to cover every possible (AFAIK) node which can meaningfully have a Type.  Type information is being rolled up the AST to the top most nodes, based on expression and operator rules defined in the JLS.
The ternary operator, and PrimaryExpressions are not yet done.  This feature will not work very well until PrimaryExpression are handled, as this node appears very often.  A failure at this level could stop the progress of the rollup, seriously limiting utility.

Also, some Parser changes are needed to (1) annotate literals with their type, and (2) know the dimensionality of array types.

Yeah, need to write a boatload of test cases too. :)


git-svn-id: https://pmd.svn.sourceforge.net/svnroot/pmd/trunk@5860 51baf565-9d33-0410-a72c-fc3788e3496d