Upgraded documentation and code of the new maxProblemCount attribute in PmdAntTask:
failOnViolation == true + problemCount not setted => fail if one violation or more is reported.
failOnViolation == true + problemCount is setted => fail if the problemCount threshold is reached
failOnViolation not setted + problemCount is setted => fail if the problemCount threshold
failOnViolation to false + problemCount is setted => depends on the order ant to the setting, most likely going not to fail.

I think this is the more appropriate solution. maxProblemCount is no longer tied up to the failOnViolation parameter. 

When a use maxProblemCount, it will automaticly set failOnViolation to true.

git-svn-id: https://pmd.svn.sourceforge.net/svnroot/pmd/trunk@5901 51baf565-9d33-0410-a72c-fc3788e3496d