Add base classes for Parser implementations to eliminate duplicate code.  Also add a TokenManager wrapper interface and implementations for each Parser.
Ultimately this stuff needs to be moved into Language specific packages.  I expect further effort will need to be done to generalize anything and everything related to Tokenizers, Parsers, ASTs, etc.  We should explore if there's a way to get JavaCC to use actual interfaces for things.  It seems to be generating duplicate one-off implementations, but does not implement a standard set of interfaces.  Once we have 3+ JavaCC implementations in PMD, we really need a way to make them appear identical to all the PMD internals.  Settling on JavaCC as THE standard for PMD seems like the easiest choice for now, we can mix things up in the future once we get it working for the simplest case.

git-svn-id: https://pmd.svn.sourceforge.net/svnroot/pmd/trunk@5931 51baf565-9d33-0410-a72c-fc3788e3496d