Start of the cleaning of pmd build process.
I added a subdirectory 'tool' with the small pom.xml required to build and package the java code source for a small artifact called 'pmd-build'. This jar contains most of the 'logic' specific to the pmd-build. This code is basicly and java port from the old jelly script (when xslt couldn't do the trick).

The source code is not really clean, i'll propably improve it in next weeks. 

Note: some other commits to do in the next minutes, please be patient, let's no try to see to what point svn is transactionnal ! :)

git-svn-id: https://pmd.svn.sourceforge.net/svnroot/pmd/trunk@6025 51baf565-9d33-0410-a72c-fc3788e3496d