Adding the generation of pmd-test.jar to the target 'jar'. It's a quick fix.
Also note that i removed the following line from the jar 
creation:
excludes="test/net/sourceforge/pmd/testframework/**/**" 
this lead to exclude the testframework itself, which lead
to a pretty useless jar.

I didn't remove the pmd-test jar creation from the 
retroweaver task neither remove this exclusion. Retro
Weaver part of the build is still not working on my 
laptop, so i didn't want to alter without the capability
to test my modifications.

I modify the 4.2.3 first because of this. This way, if we have to
release a 4.2.3 before 5.0, the pmd-test.jar should be
in it.


git-svn-id: https://pmd.svn.sourceforge.net/svnroot/pmd/branches/pmd/4.2.x@6139 51baf565-9d33-0410-a72c-fc3788e3496d