Add support for XPath 2.0 and XPath 1.0 compatibility using Saxon.  This is implemented using the Saxon OM classes, for good performance.  Existing PMD XPath functions are refactored to be compatible with Saxon's extension support.  The functions are exposed using 'pmd' and 'pmd-java' namespace prefixes, while these prefixes cause no issues for Jaxen, they are required by Saxon.
If you want to experiment using XPath 1.0 compatibility mode, simply change the default index on the 'version' enum property in XPathRule to a 1.  All Rules work properly in this mode using Saxon.  RuleChain is not in use, so performance is not currently as good as Jaxen, but it's not horrendous either.


git-svn-id: https://pmd.svn.sourceforge.net/svnroot/pmd/trunk@6491 51baf565-9d33-0410-a72c-fc3788e3496d