Remove XPath 2.0 for XML rules, for now.  I thought it would be a nice test case, but it's broken on Java 1.4.
I thought I had verified Saxon worked on Java 1.4, but apparently it cannot due in part to dependencies on the DefaultHandler2 class which is only available in Java 1.5 and later (or Java 1.4 with JAXP extension).  I think we'll just have to document this limitation for folks wanting to continue to use Java 1.4.

git-svn-id: https://pmd.svn.sourceforge.net/svnroot/pmd/trunk@6549 51baf565-9d33-0410-a72c-fc3788e3496d