Rework Renderer to support standardized option passing, as well as a more rigorous definition of those options.  This approach allows for a RendererFactory to create Renderer instances, as well as a means to programmatically output details on all available Renderers.  Both the PMD command line and Ant tasks now take advantage of the RendererFactory, and have been enhanced to support generic report option passing (ie. -parameter and <param> respectively).  Removed obsoleted Renderer specific options from command line and Ant task (e.g. linkPrefix, etc).  Finally, some Renderers were given more appropriate names.  We should consider dropping some of these goofy Renderers out of the tree (there's about 4-5 Renderers for HTML!).
Added missing -showsuppresse option for PMD command line.  The default is the same as the Ant task (false), which is a change from prior PMD behavior, but I think this is justified for consistency's sake and the fact that most people likely use the Ant task anyway.

While working on CommandLineOptions, the argument processed code was cleaned up a bit so that missing parameters for options will now give better error messages.  This isn't to say there's not work that could still be done in this area, as the code is still rather too loosey-goosey for my tastes.


git-svn-id: https://pmd.svn.sourceforge.net/svnroot/pmd/trunk@6561 51baf565-9d33-0410-a72c-fc3788e3496d