Rename SourceFileSelector to LanguageFilenameFilter.  Delete SourceFileOrDirectoryFilter.  Move DataSource finding code from the PMD class into FileUtil.
Overall, nice bit of cleanup and removal of code from the PMD class into a reusable utility (used by PMD and the Benchmark program).

Added a FUTURE marker as we still need to figure out a way to deal with excluding certain directories, the hard-coded "SCCS" is still in place.  Thankfully this only applies to command line usages, and not Ant.


git-svn-id: https://pmd.svn.sourceforge.net/svnroot/pmd/trunk@6588 51baf565-9d33-0410-a72c-fc3788e3496d