Completed refactoring of PMD configuration state into the Configuration class.  Both CommandLineOptions and the AntTask refactored to initialize a Configuration instance as needed.  This results in much improved defaulting of common settings and behavior between the command line and Ant task.
Remove many redundant "process" methods from the PMD class.  Use of a Configuration object removes the need for so any variants.

Renamed the 'cpus' option to 'threads', for both command line and Ant.

Changed the 'lang' option to 'version', for both command line and Ant.  The option can be specified repeated to change what LanguageVersion to use for any given Language.  For Ant, use <version> child element.  Reworked the way PMD processes to let the Languages of the Rules in the RuleSets drive the processing, instead of the previous single language.

It should now be technically possible to process multiple Languages in the same RuleSet on a list of mixed Language files.  However, I still need to create tests for this scenario to verify it is working as expected.


git-svn-id: https://pmd.svn.sourceforge.net/svnroot/pmd/trunk@6615 51baf565-9d33-0410-a72c-fc3788e3496d