Refactor SimpleRuleSetNameMapper to understand the Language prefix.  For example 'basic' is now 'java-basic'.
Refactor all RuleSet XMLs into a Language subdirectory.  The 'rulesets.properties' file can be found in each directory.  The RuleSetFactory when asking for all registered RuleSets will use the 'rulesets.properties' file for each Language with Rule support.

All regression tests updated to use Language scoped short names, instead of old style.

Renamed the new 'useless' ruleset to 'unnecessary', as it is a more mild term.

Various parts of the documentation and the release scripts will need updating to account for the new treatment of 'rulesets.properties', and ruleset naming.  This can be done later, when all documentation is given a thorough thrashing. ;)


git-svn-id: https://pmd.svn.sourceforge.net/svnroot/pmd/trunk@6630 51baf565-9d33-0410-a72c-fc3788e3496d