Add support for ECMAScript.  Use Rhino as a Parser and AST producer.  Added 3 initial Rules as proof of concept.
The adaption of the Rhino AST to the PMD AST is done via a custom PMD AST class per Rhino AST class.  This PMD AST class exposes those parts of the underlying Rhino AST node as needed, and also provides intelligent access to the children of the PMD AST node, similar as has been on for the PMD Java AST and in Rhino itself.

The EcmascriptTreeBuilder provides a model for how to efficiently traverse a non-PMD AST, and convert a corresponding PMD AST.  This happens to be _much_ faster than any dynamic proxy approach could be (already tried a rough cut, this approach is over 10x faster).  In the case of Rhino, we leverage the existing NodeVisitor API provided on the Rhino AST.

This approach relies heavily on the Rhino AST model, which is not "published" as far as I can tell, and is significantly changed in the CVS trunk over the most recently released version of Rhino.  Dealing with changes in the Rhino model will one of the downsides to leveraging Rhino, but upside being we have a fully functional Language available with a little more than a day's worth of coding.


git-svn-id: https://pmd.svn.sourceforge.net/svnroot/pmd/trunk@6707 51baf565-9d33-0410-a72c-fc3788e3496d