Open things up to support files other than .java files.  This includes: 1) Allow the PMD context menu on any File type, not just *.java. 2) Modify some places which check for a .java extension to _not_ do so. 3) Modify some places to check with the PMDEngine whether a File "applies" (e.g. is it for a Language which is understood by PMD and are there any Rules in the RuleSet for that Language).
Expand the Rule table to have a Language column in the 1st column.  The table setup in this plugin/SWT still *sucks* as bad I remember it being the last time I added columns.

With these changes I can get both JS and XML Rules to work.  Violations show up, and for JS, I can be brought to the proper line number in the code!

Also fixed and issue with this plugin looking for a "maxViolations" property on the Rules, which given PMD 5.0 changes to require a propery to exist to ask for a value... needed some tweaking to first check for the value on the Rule.  More verbose, but a safer API than the Wild West style of pre-5.0.

The Rule Editor needs some work to account for the new structure of Rules in PMD 5.0.  Currently any Rule it creates causes things to blow up due to missing Language.  Easy enough to fix, but it's just the 1st of several things needing to be corrected.




git-svn-id: https://pmd.svn.sourceforge.net/svnroot/pmd/trunk@6742 51baf565-9d33-0410-a72c-fc3788e3496d