Working on bug : https://sourceforge.net/tracker2/?func=detail&amp;aid=2410201&amp;group_id=56262&amp;atid=479921
I reproduced the issue and trace the problem back to the symbol table (More exactly a method of the NameOccurences class). I also added a test to reproduced the issue here. (Also added a little javadoc).

I'm not really although how to "assert" results in this test, not really sure how the symbol table should be expected to behave here.

No more time right now for this, i'll try see through this week. 

Note: I'll patch trunk when I'll have fix for good this issue.

git-svn-id: https://pmd.svn.sourceforge.net/svnroot/pmd/branches/pmd/4.2.x@6759 51baf565-9d33-0410-a72c-fc3788e3496d