remove dependency to asm in project.xml
As it seems that ASM is no longer available in Maven 1 repository, this dependency leads to
a build failure. As Maven 1 is ONLY used to generate the xdocs documenation, and has been removed
from the trunk, it does not seems relevant to try to fix this. Removing the depedency does allow
one to build the documentation, which what we need here...

git-svn-id: https://pmd.svn.sourceforge.net/svnroot/pmd/branches/pmd/4.2.x@7246 51baf565-9d33-0410-a72c-fc3788e3496d