pmd (build): Refactor pmd-build code
The tools subproject was created over 2 years ago to simplify the build
process and enable both Maven 2 and Ant to easily do what maven 1 was
actually doing for us. The present refactoring has greatly clarified the
code and allow the build to process rulesets's directory recursively, in
order to cope with new rulesets directory layout.

Note: as we are now moving forward with using only mvn as our build system
this subproject will probably wrap into a regular mvn-plugin, and therefore
should be moved from the pmd/ directory to the root of our svn. Maven plugin
integration should reduced greatly the complexity of the pom.xml and remove
the pmd-build-*.jar from the repository... (however, it remains to be done).

git-svn-id: https://pmd.svn.sourceforge.net/svnroot/pmd/trunk@7519 51baf565-9d33-0410-a72c-fc3788e3496d