pmd (build): Fix bug in site's menu + logging
Fix a bug in the website generation leading all entries in the rulesets navigation
menu to be duplicated.

Also introduce Java Logging, instead of System.out, as a logger for the tools's
subproject. (Rationale behing that is - I hope, an easier integration with Maven later
on without making this code dependant of any librairies outside the JVM).

git-svn-id: https://pmd.svn.sourceforge.net/svnroot/pmd/trunk@7528 51baf565-9d33-0410-a72c-fc3788e3496d