pmd: No documentation for code based rule descriptor.
Currently it's not possible to generate
documentation for a rule with a code based
descriptor - short of generating an XML file from
it and process it through the current set of XSL
transformation we use to produce the website.

While it's a nice idea, I don't think this is
eleganlty doable and I do think it's something
"people are desperatly" waiting for. So, as I
don't want this item to block the release, I'll
remove it...

git-svn-id: https://pmd.svn.sourceforge.net/svnroot/pmd/trunk@7542 51baf565-9d33-0410-a72c-fc3788e3496d