cpd: Rework XML renderer (fix: bug 1435751)
CPD XMLRenderer was still manipulating strings and strings buffer
which lead to countless issues with proper encoding. It now uses
the appropriate XML API and relies ONLY on the defined system
variable file.encoding to set the encoding.

If one uses Ant or CLI option to change the encoding, the value of
file.encoding will then adapt. It may sounds a little bit extreme
but it should ensure consistency of all the files...

git-svn-id: https://pmd.svn.sourceforge.net/svnroot/pmd/trunk@7547 51baf565-9d33-0410-a72c-fc3788e3496d