Refactor RuleViolationFactory
In order to ease the integration of new language,
I've refactor a couple of Abstract class, that was
in fact cocncrete (ie no abstract methods) into
parametizable (if this is a word) classes. Idea
being that other languages can simply reuse the
class instead of having to create a class just
to make the abstract class concrete.

git-svn-id: https://pmd.svn.sourceforge.net/svnroot/pmd/trunk@7627 51baf565-9d33-0410-a72c-fc3788e3496d