pmd-eclipse: Support workspace-relative path in ruleSetFile property
When using multiple projects that want to share the rule set definitions
in a single file, it's currently only possible to reference the shared
file using an absolute path. The downside of the absolute path is that
it is different for each developer, so the resulting .pmd file can not
be committed in the repository.

To fix this, this change also tries to resolve the path against the
workspace root, so e.g. "/foo/.ruleset" can be used to refer to the
.ruleset file in the foo project.

Such a path can be ambiguous to an absolute file system path, but in
practice it should not be a problem. An alternative solution would be to
support variables (as known from Eclipse run configurations) such as
"${project_loc:foo}" in the path.