Some code enhancements:
PMD.java. splited the imports to avoid check style message: Using the '.*' form of import should be avoided - net.sourceforge.pmd.lang.*.
AbstractNodeInfo now has all the overrides
String utils now has the EMPTY_STRING variable declared as private and a getter method to access it. Refactored isSame to avoid using     @SuppressWarnings("PMD.CompareObjectsWithEquals")
DFAPanel uses the getter for empty strings. Also deleted comments that seems to be old code and addd the override to valueChanged
In DBMSMetadadataTest we were passing a parameter to  System.out.format but it hasn't the place holder for the parameter