PR18758: Fixed NullPointerException in ConstructorCallsOverridableMethod.
This problem does not occur anymore in pmd/master. However, by looking at the diff it seems that this is not due to a fix in ConstructorCallsOverridableMethodRule.java iteself. I think it has something to do with fixes in the Java parser (Java.jjt) but I was not able to isolate the fix. Therefore, solved it by a simple null guard.