POC - add a PmdThreadContextHolder to have access to the report anywhere.
API is ugly and needs to be defined yet. The problem should be reported
as a configuration error rather than a processing error. But a
configuration error needs a rule and is unfortunately currently
ignored by the maven-pmd-plugin... (it doesn't end up in the report).

But in general, it works - the missing classes end up as processing
errors in the final report.

I probably missed some cases where we should report missing classes
or reporting the cases multiple times...
In the end, we should probably deduplicate and report each missing
class only once.