Perform preliminary ambig counting
13% of all expressions are ambiguous names
55% percent of those ambiguous names can be resolved to values declared in the ACU
20% other percent can be resolved to an imported type name
In total, only 3% of all expressions are ambiguous without any classpath support

Those measurements are skewed by the shitty type resolving of TypeSet,
which throws away a resolved FQCN when the class is not on the classpath
Probably, many more names can be resolved to types.

The remaining semantically ambiguous names would be all resolved when the
auxclasspath is complete.

When it isn't, a naming convention and other heuristics could be used to remove
the remaining ambiguous names with minimal risk of error.
Rules like UnnecessaryQualifiedName already use heuristics themselves
and their level of precision is very good. So if that's acceptable, why not share
that analysis with everything and have a more semantic AST?