Remove findLanguageVersionByTerseName
This method was stupid, most usages actually concatenate two strings
to pass as an argument. The only place this is used appropriately is
when parsing the sourceType of rule tests, with the caveat that the
source may be a different language than the rule, which makes no sense
at all.