[core] Fix XPath node compareOrder
In Saxon, NodeInfo.compareOrder is used to determine, whether a
node is already contained in a union set or not. For different nodes
it must return a value != 0. It should consider the position of the
nodes in document order. The new implementation uses now the
position of the nodes (begin line, begin column) to determine the order.
It might fail, if the position of the nodes are not correct.