Cleanup PMD.EOL in tests
Instead, normalize line endings in parse method.
This makes sure test code snippets are portable
and readable. For now they're platform-specific,
and assume that the resource files are saved with
the platform-specific line separator. This may not
be the case, as eg a git file may be checked out
with CRLF on a unix system.