Added column information to CPD internals.
Duplications are registered in CPD by storing a Mark. This Mark contains a TokenEntry for where the duplication starts, as well as a line count and the source code that is duplicated.

This change adds a beginColumn and endColumn field to each TokenEntry. These are optional fields that can be left empty. Storing these allows us to pinpoint the column position of each token.

In addition, an additional TokenEntry is added to the Mark to indicate where the duplication ends. This can then be used to determine the location of the entire duplication; it starts where the first token starts and ends where the last token ends.