Fix getSlice
The method path for when the
code is not loaded yet was incorrect.

This was hidden by the fact that tests
use a dummy tokenizer *before* cutting
out a slice, so that the code is in the
soft reference, and that code path is
never taken in the relevant tests.