Ignore threading
We consider that reaching field defs are observable
at the point of the leak. But we assume the ctor
executes atomically, so following def chains work
normally.

Maybe, if the field is volatile, we should be more
conservative and assume that all its assignments after
any leak are observable. This looks very rare though