Mention pmd-java deps explicitly
They're transitively gotten from
pmd-core, but for some use cases
this is nicer

eg `mvn install -pl pmd-java`, given
that pmd-core was already installed
locally, should succeed. If the dependencies
are not mentioned explicitly, for some
reason this fails.

Before #2614, pmd-core was shaded with
its dependencies and this wasn't necessary