Fix stackoverflow in spring
The problem was that unresolved types had an unresolved type as
a superclass. So their ancestors were basically an infinite
chain (which is lazily resolved though). When calling isSubtype,
at some point we recurse into ancestors and trigger the infinite
recursion.

With enough memory this manifests as a StackOverflowError. Before
I upgraded my RAM though, this threw an OutOfMemoryError (because
of the locals on the stack).