Fix importing bug
Shows up in java.base

The problem is importing a member declared in the file will try to recurse
into the superclass to find members with the same name.
But the superclass node has not been disambiguated yet,
and cannot be, as scopes are not properly built.
We just load static imports lazily now.