[ci] Improve pmd regression tester call
When executing regression tester against master, we
don't actually run PMD again, but we use the result from
the previous run. There only the modified rules have been
executed, so we should use the same rules for comparison
against master.
This should result in smaller reports, since the comparison
on master reported all the not executed rules as missing
violations.