[core] Fix LanguageVersion compareTo
This didn't work with java 9 compared to java 10,
as the full name was compared ("Java 9" vs. "Java 10")
as a string rather than using version numbers.

Now the logic is much simpler: The versions defined
for a language are ordered, new versions are always added
at the end.

Also backports the useful method "compareToVersion" from pmd 7
and use this in the rules BigIntegerInstatiation and
UnnecessaryWrapperObjectCreation.