Issue 3569 - Based on great feedback from Clément Fournier, decided to go a different route and add the exact same configuration properties that the analogous Java rule offers, specifically 'allowCommentedBlocks' and 'allowExceptionNameRegex', both with the exact same defaults as the Java rule, i.e., 'false' and '(expected|ignored)' respectively.
Adding 'allowExceptionNameRegex' was quite straightforward. However, 'allowCommentedBlocks' was a bit more challenging because there was no notion of whether a block contained comments. I had to augment the lexer/parser/tree builder logic to capture all comment tokens and then use them to denote whether a catch block contained comments of any sort. The way I did this with a new intermediate abstract base node class for comment containers, we could denote the same for the other Apex block node types. In the spirit of keeping these changes narrow in scope, though, I only changed the inheritance hierarchy for ASTCatchBlockStatement. If/when we need to be able to ask the same question about other block types, it should be trivial to update them to support it.

Test cases have been updated accordingly.