Fix #3763 - ambiguity error
Problem was that the header scopes looked like this:

   IMPORTED -> EMPTY -> JAVA_LANG

Where IMPORTED and JAVA_LANG both can resolve the same
name NAME. The empty symtable is administrative, it's
a shadow barrier. With the previous definition of a
 shadow barrier, the barriers were positioned this way:

   IMPORTED -> EMPTY] -> JAVA_LANG]

IMPORTED asks for NAME and finds something. It's not
a shadow barrier so it asks EMPTY for NAME. EMPTY
does not know NAME so it asks its parent, ignoring that
it itself is a shadow barrier.

Now we don't call `ShadowChain::resolve` to collect
shadowed names to avoid this situation.