Add simplelogger.properties in pmd-test
This provides a default configuration that will be used
when resetting the logger configuration in CLITests or
PMDTaskTest. Without this, slf4j-simple will keep
the last configured default level. Now it will use the
default level configured in simplelogger.

Since pmd-test is not part of the distribution, this
configuration file has no effect for the binaries.