Internalize logger
I don't think it's worth removing it in master...
Traces could admittedly be reported on a jutil.Logger,
but this makes the code unreadable and it's not
clear why we would do that. Other things like
warnings and errors will be reported through
this interface in pmd 7, so it also doesn't
make sense to update their call sites.