Fix OutOfMemory in tests
This did not show up earlier because
we used soft class references, which
reduces the memory footprint of unused
TypeSystems. However, they still remain
in old gen and are a large memory hog.

Since e97e2cd we don't use SoftReference
anymore as parsing the same class several
times may have unintended side-effects.
This revealed the out of memory error.

This will be fixed when we have proper
language lifecycle and resources like the
classloader can be reclaimed later.