[java] Ignore synthetic methods and fields in tests
When running PMD build for sonar and coveralls, the jacoco agent will insert a synthetic method "$jacocoInit()" and synthetic field "$jacocoData".
Both are not present in the class files on disk, which
are used for the asm-based symbol analyzer.

For now, all synthetic methods and fields are ignored
and not just the two jacoco specific ones.