[java] SwitchDensityRule - use common report level property for "minimum"
While we calculate the ratio between number of statements and number of labels,
we shouldn't pretend that this is an exact rule - it's anyway a best guess,
that high switch density is a problem. I don't think, it makes a big difference
between 10.2 and 10. You e.g. want to report, when the density is bigger than 10
or bigger than 5. We don't need the report level accuracy here.

Refs #3123