[core] Support loading classes from runtime images
That's one missing piece to support running PMD on a different JRE
than the one used for analyzing. For Java up to version 8, one can put
lib/rt.jar on the auxclasspath. But since Java 9, the runtime classes
are stored in runtime images, that need to be loaded through the
jrt:/ file system.

Since the jrt-URLs are not connected anymore to a specific runtime
image, AsmSymbolResolver needs to use directly the streams instead
of URLs.

The ClasspathClassLoader is basically disabled to actually load
classes for reflection (#loadClass). In PMD 7, we shouldn't use that
anymore.

See also https://openjdk.org/jeps/220