Fix line/col numbers in TokenMgrError
Ref #4635

Col number of zero may happen when the error is reported
on a newline. In the rest of PMD a newline is considered
to sit at the last index of the previous line. So the index
is now the first character of the next line, which is slightly
wrong but that doesn't matter much.