[core] Remove support for deprecated rule set references notation
The old notation <lang>-<ruleset> is not supported anymore. It is
now interpreted as a ruleset reference without a ruleset and just
referencing a single rule. Also the release number notation is
not supported anymore.

Since RuleSetReferenceId is Internal+Deprecated, no API changes.
Clarified External/Internal: A RuleSetReference is either absolute
(RuleSet is known) or relative (RuleSet is not known).

Fixes #4313