(PoC) [java] Verify JDK API Version
This is just a PoC to get the idea out. We should implement something
like this to warn users, if they execute PMD with incorrect java
versions. Using the wrong java version on the auxclasspath leads
to issues like #4620 - the Java API changed between and
typeresolution works as designed. But with the wrong Java Runtime
on the auxclasspath, false positives are detected.

You can specify the java runtime on the auxclasspath via
CLI as described on https://docs.pmd-code.org/latest/pmd_languages_java.html#providing-the-auxiliary-classpath .
But if no java runtime is provided, we fall back to the
runtime java version, which might or might not be correct.

Note: For maven, there is currently no official way to configure
the java runtime onto the auxclasspath. You can workaround by
adding a system scoped dependency to your project, e.g.

```xml
        <dependency>
            <groupId>java8</groupId>
            <artifactId>java8-rt</artifactId>
            <version>8</version>
            <scope>system</scope>
            <systemPath>/path/to/jdk-8/jre/lib/rt.jar</systemPath>
        </dependency>
```

Since maven adds all project dependencies to the auxclasspath when
executing PMD, this adds the java 8 runtime and false positives
such as #4620 disappear.

In theory, you could also use maven toolchains to execute PMD
with the correct java version.

At the beginning I think, we should issue a warning, if we
detect a mismatch between the java language version and the java
version, that we resolve from the auxclasspath. Note - the
warning I added in this PR only appears in DEBUG mode.
And I assume (needs to be verified) that we could do this check
once at the beginning and don't need to issue a warning for
every file...

Once the warning is out there and we provided enough documentation
around how to resolve this warning, we could make this into a
fatal error, aborting the PMD analysis. We might even think
about not falling back to the runtime classpath for type resolution,
forcing users to always configure the intended java version
explicitly.