[ci] New optimized workflow for pull requests
The goal is to have faster feedback for PRs
and better visibility which part of the workflow part failed.

- First a fast compile job. The result of this job is shared via
  compile-artifact with the following jobs.
- Then in parallel a bunch of other jobs
  - verify-complete: runs a complete `./mvnw verify` with all tests,
    code checks, etc. This is only run on linux.
  - verify-unittests: just runs the unit tests on windows and macos
  - dogfood: run PMD on PMD with the latest changes from the pull
    request
  - regressiontester: runs the pmdtester to produce the regression
    report

Since most jobs can be run in parallel and not sequentially, the
results of the PR build should be available faster.

The concurrency control makes sure, that we cancel any in-progress
jobs for the current pull request to avoid unnecessary builds.
Only the latest build matters.