[ci] build: Add pmdtester_start.sh to run regression tester
This is now more flexible and supporting to fetch additional commits
not only from PRs but also from branches.
This means, a push on a branch other than main will create a report
diffing the branch against main, just like a PR. This should be
handy for forks/local branch developments.

And a push on the main branch will create a regression report
by looking ath the pushed commits only
(using github.event.push.before as base). This should create a report
showing the differencs caused by the push. Note: Since we use
the last created regression tester baseline, this might not be
accurate, though.

We need to really use "main" as the base branch, as
pmd-regression-tester will download the baseline using this
base branch name as the filename.