[ci] build: better cross-os caching for local maven repo
"~" would be expanded to different absolute paths on windows/mac
and thus the maven repo is not really reused. Now maven is using
the local directory ".m2/repository" for the repo cache. This
directory is within the current workspace of the runner and
a relative path is used for the cache.

Now we also need to exclude ".m2" from the src distribution.