[ci] Reuse GitHub Pre-Releases
By reusing an already existing pre-release, no notifications should
be sent when updating the release. Previously, we deleted the old
pre-release and created a new pre-release, which resulted in a
notification about a new release for every build from branch main.

Reusing the pre-release is not straight forward: releases are
coupled to tags. Deleting a tag for an already existing release
changes that release to "draft" which then needs to be published
again. But we don't want to publish the release again to avoid
unnecessary notifications about a pre-release. That's why a
new temporary tag is created, the existing pre-release is moved
onto this temporary tag and then the old tag is removed and
recreated to point to the tip of the main branch. This effectively
moves the release tag for the pre-release forward. After that
the release is updated with the new information.

The build-release.yml and git-repo-sync.yml workflows are updated
to ignore these "SNAPSHOT" release tags.